+++ src/config/app.php
@@ -3,8 +3,8 @@
 return [
     'id' => 'CraftCMS',
     'name' => 'Craft CMS',
-    'version' => '3.6.5.1',
-    'schemaVersion' => '3.6.2',
+    'version' => '3.6.6',
+    'schemaVersion' => '3.6.4',
     'minVersionRequired' => '2.6.2788',
     'basePath' => dirname(__DIR__), // Defines the @app alias
     'runtimePath' => '@storage/runtime', // Defines the @runtime alias