+++ src/elements/db/ElementQueryInterface.php
@@ -153,6 +153,8 @@ interface ElementQueryInterface extends QueryInterface, ArrayAccess, Arrayable, 
      * | - | -
      * | `1` | for the {element} with an ID of 1.
      * | a [[{element-class}]] object | for the {element} represented by the object.
+     * | `'*'` | for any {element}
+     * | `false` | that aren’t associated with a published {element}
      *
      * ---
      *
@@ -170,7 +172,7 @@ interface ElementQueryInterface extends QueryInterface, ArrayAccess, Arrayable, 
      *     ->all();
      * ```
      *
-     * @param int|ElementInterface|null $value The property value
+     * @param int|ElementInterface|string|false|null $value The property value
      * @return static self reference
      * @since 3.2.0
      */
@@ -207,6 +209,33 @@ interface ElementQueryInterface extends QueryInterface, ArrayAccess, Arrayable, 
      * @since 3.2.0
      */
     public function draftCreator($value);
+
+    /**
+     * Narrows the query results to only unpublished drafts which have been saved after initial creation.
+     *
+     * ---
+     *
+     * ```twig
+     * {# Fetch saved, unpublished draft {elements} #}
+     * {% set {elements-var} = {twig-function}
+     *     .draftOf(false)
+     *     .savedDraftsOnly()
+     *     .all() %}
+     * ```
+     *
+     * ```php
+     * // Fetch saved, unpublished draft {elements}
+     * ${elements-var} = {element-class}::find()
+     *     ->draftOf(false)
+     *     ->savedDraftsOnly()
+     *     ->all();
+     * ```
+     *
+     * @param bool $value The property value (defaults to true)
+     * @return static self reference
+     * @since 3.6.6
+     */
+    public function savedDraftsOnly(bool $value = true);
 
     /**
      * Narrows the query results to only revision {elements}.