+++ src/test/TestSetup.php
@@ -490,14 +490,9 @@ class TestSetup
             'password' => 'craftcms2018!!',
             'email' => 'support@craftcms.com',
             'site' => $site,
-            'applyProjectConfigYaml' => false,
         ]);
 
         $migration->safeUp();
-
-        if ($projectConfig) {
-            Craft::$app->getProjectConfig()->applyYamlChanges();
-        }
     }
 
     /**